requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv
prefect
slack_sdk
